String类和Static关键字，继承和抽象 --bylM
