+++ src/cn/lm/demo/继承抽象/Fu.java
@@ -1,0 +1,6 @@
+package cn.lm.demo.继承抽象;
+
+public class Fu {
+    //Fu中的成员变量
+    int num = 5;
+}
