+++ src/cn/lm/demo/继承抽象/TestDemo01.java
@@ -1,0 +1,12 @@
+package cn.lm.demo.继承抽象;
+
+public class TestDemo01 {
+    public static void main(String[] args) {
+        //创建子类对象
+        Zi z = new Zi();
+        //调用子类中的show方法
+        z.show();
+    }
+
+
+}
